AddComponentTag uses default component font color
`AddComponentTag` and `AddExternalComponentTag` should use the `$COMPONENT_FONT_COLOR` by default instead of the `$ELEMENT_FONT_COLOR`. Without this change, a component with a tag that is defined by `AddComponentTag` will have a different font color even if $fontColor is not passed in, because the font color used by default components is different than the default font color set by `AddComponentTag`.